Tidy up featured albums
